fixed errors in dummy data
